implemented usage of recipe read by rtde client in the parser and the rtde data package

+++ src/rtde/control_package_pause.cpp
@@ -40,7 +40,7 @@ bool ControlPackagePause::parseWith(comm::BinParser& bp)
 std::string ControlPackagePause::toString() const
 {
   std::stringstream ss;
-  ss << "accepted: " << accepted_;
+  ss << "accepted: " << static_cast<int>(accepted_);
 
   return ss.str();
 }
